Revert "Merge pull request #57 from EFUB-TEAM4/feature/weather"
This reverts commit 937c60e7254e58d59e8aa79bc17781342762413f.
