+++ src/main/java/efub/team4/backend_eweather/domain/bear/controller/BearController.java
@@ -2,10 +2,8 @@ package efub.team4.backend_eweather.domain.bear.controller;
 
 import efub.team4.backend_eweather.domain.bear.dto.BearImageResponseDto;
 import efub.team4.backend_eweather.domain.bear.service.BearService;
-import efub.team4.backend_eweather.domain.weather.service.OpenWeatherAPI;
 import lombok.RequiredArgsConstructor;
 import org.json.simple.parser.ParseException;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -16,16 +14,11 @@ import java.io.IOException;
 @RestController
 @RequestMapping("/api/v1/bear")
 public class BearController {
-
-    @Autowired
     private final BearService bearService;
-
-    @Autowired
-    private final OpenWeatherAPI openWeatherAPI;
 
     @GetMapping
     public BearImageResponseDto loadBearImage() throws IOException, ParseException {
-        return bearService.findBearImage(openWeatherAPI.findBearInfo());
+        return bearService.findBearImage();
     }
 
 }
