+++ src/main/java/efub/team4/backend_eweather/domain/weather/controller/OpenWeatherApiController.java
@@ -15,6 +15,7 @@ import java.util.List;
 @RequestMapping("api/v1/weather")
 public class OpenWeatherApiController {
 
+    @Autowired
     private final OpenWeatherAPI openWeatherAPI = new OpenWeatherAPI();
 
     @GetMapping
